+++ db/seeder/src/seed.rs
@@ -4,7 +4,7 @@ use web_server::domain::SafePassword;
 
 const TEST_PASSWORD: &str = "Test123!";
 
-pub async fn seed(pool: PgPool, count: u32) -> anyhow::Result<()> {
+pub async fn seed(pool: PgPool, count: usize) -> anyhow::Result<()> {
     let passhash = SafePassword::parse(TEST_PASSWORD.to_string())
         .unwrap()
         .hash()
